move fuzzy match utils to core
